Date deltas can now be negative
Past Humberto forgot that subtractions could be negative.

+++ trabalho-pratico/src/utils/date.c
@@ -170,14 +170,14 @@ void date_sprintf(char *output, date_t date) {
             date_union.fields.day);
 }
 
-uint64_t date_diff(date_t a, date_t b) {
+int64_t date_diff(date_t a, date_t b) {
     date_union_helper_t a_union = {.date = a};
     date_union_helper_t b_union = {.date = b};
 
-    uint64_t a_days =
-        (uint64_t) a_union.fields.year * 12 * 31 + a_union.fields.month * 31 + a_union.fields.day;
-    uint64_t b_days =
-        (uint64_t) b_union.fields.year * 12 * 31 + b_union.fields.month * 31 + b_union.fields.day;
+    int64_t a_days = (int64_t) a_union.fields.year * 12 * 31 + (int64_t) a_union.fields.month * 31 +
+                     (int64_t) a_union.fields.day;
+    int64_t b_days = (int64_t) b_union.fields.year * 12 * 31 + (int64_t) b_union.fields.month * 31 +
+                     (int64_t) b_union.fields.day;
 
     return a_days - b_days;
 }
